Split a test file
